% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{accessibility_report}
\alias{accessibility_report}
\title{Accessibility report}
\usage{
accessibility_report(
  file,
  profile = "ua1",
  output_file = tempfile(fileext = ".html"),
  open = TRUE
)
}
\arguments{
\item{file}{PDF file to check.}

\item{profile}{The validation profile to use. Default to \code{"ua1"}.}

\item{output_file}{Path for the HTML report. If \code{NULL}, creates a temp file.}

\item{open}{Whether to automatically open the report in browser. Default \code{TRUE}.}
}
\value{
Path to the generated HTML report (invisibly)
}
\description{
Generates an HTML report on accessibility for a given PDF.
}
