% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{info}
\alias{info}
\alias{get_total_failed_checks}
\alias{get_total_failed_rules}
\alias{get_verapdf_version}
\title{Retrieve specific accessibility info}
\usage{
get_total_failed_checks(x)

get_total_failed_rules(x)

get_verapdf_version(x)
}
\arguments{
\item{x}{Output from \code{verapdf()}}
}
\description{
Utility functions to retrieve specific informations
about PDF accessibility.
}
\examples{
\dontrun{
verapdf("inst/pdf/not-compliant-1.pdf") |>
  get_total_failed_checks()

verapdf("inst/pdf/not-compliant-1.pdf") |>
  get_total_failed_rules()

verapdf("inst/pdf/not-compliant-1.pdf") |>
  get_verapdf_version()
}

}
