% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verapdf-cli.R
\name{verapdf}
\alias{verapdf}
\title{Call verapdf CLI}
\usage{
verapdf(
  file,
  write_to = NULL,
  format = c("json", "xml"),
  profile = c("ua1", "ua2", "1a", "1b", "2a", "2b", "2u", "3a", "3b", "3u", "4", "4f",
    "4e")
)
}
\arguments{
\item{file}{PDF file to check.}

\item{write_to}{Path to output file. If \code{NULL}, does not write.}

\item{format}{Output format. Default to \code{"json"}.}

\item{profile}{The validation profile to use. Default to \code{"ua1"} (recommended).}
}
\value{
output from the CLI
}
\description{
Utility to call verapdf command line interface. It requires
the \code{verapdf} CLI to be on PATH.
}
