% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{get_total_failed_checks}
\alias{get_total_failed_rules}
\alias{get_verapdf_version}
\title{Retrieve specific accessibility info}
\usage{
get_total_failed_checks(x)

get_total_failed_rules(x)

get_verapdf_version(x)
}
\arguments{
\item{x}{Output from \code{verapdf()}}
}
\description{
Utility functions to retrieve specific informations
about PDF accessibility.
}
\note{
It is important to understand the difference between
rules and checks. A rule that has been failed may
be "Missing alternative text on images", and each rule
may have several checks that have been failed.
For example, if the rule "Missing alt text on images" occurs
10 times in a given PDF, that gives us 10 failed checks
for that rule.
}
\examples{
pdf_file <- system.file("pdf", "not-compliant-1.pdf", package = "pdfcheck")

verapdf(pdf_file) |>
  get_total_failed_checks()

verapdf(pdf_file) |>
  get_total_failed_rules()

verapdf(pdf_file) |>
  get_verapdf_version()
}
